+++ app/components/Footer.vue
@@ -1,13 +1,35 @@
 <template>
-  <div>
-    
+  <div class="flex justify-center bg-tertiary py-10">
+    <div class="flex flex-col bg-white rounded-2xl w-[90%] lg:w-[1100px]">
+      <div class="flex flex-col text-center md:text-left md:flex-row justify-between items-center gap-6 md:gap-20 p-10">
+        <div class="flex flex-col gap-6 md:gap-4">
+          <p class="font-inter font-bold text-3xl">ROBALOXS</p>
+          <p class="font-inter">Dapatkan Robux dengan harga Murah, ter-aman dan tercepat hanya disini! Dapatkan Robux dengan harga Murah, ter-aman dan tercepat hanya disini! </p>
+          <div class="flex gap-5">
+            <a href="/" class="font-inter font-bold hover:text-black/70 transition-all duration-300">Privacy & Policy</a>
+            <a href="/" class="font-inter font-bold hover:text-black/70 transition-all duration-300">Terms of Service</a>
+          </div>
+        </div>
+        <div class="flex flex-col items-center gap-5">
+          <div class="flex flex-col gap-3 w-max">
+            <ButtonsPrimary variant="outlined" color="secondary" icon="mdi:arrow-right-circle-outline">Beli Robux</ButtonsPrimary>
+            <ButtonsPrimary variant="solid" color="primary" icon="mdi:arrow-right-circle-outline">Hubungi Kami</ButtonsPrimary>
+          </div>
+          <div class="flex gap-4">
+            <ButtonsIcon icon="mdi:facebook"/>
+            <ButtonsIcon icon="mdi:instagram"/>
+            <ButtonsIcon icon="mdi:twitter"/>
+            <ButtonsIcon icon="mdi:youtube"/>
+          </div>
+        </div>
+      </div>
+      <div class="flex flex-col md:flex-row my-5 mx-5 md:mx-10 justify-between items-center px-10 py-2 rounded-full bg-tertiary">
+        <div class="flex flex-row gap-1">
+          <p class="font-inter">© Copyright 2025</p> 
+          <span class="font-bold">ROBALOXS.</span>
+        </div>
+        <p class="font-inter">All Rights Reserved</p>
+      </div>
+    </div>
   </div>
 </template>
-
-<script lang="ts" setup>
-
-</script>
-
-<style>
-
-</style>