+++ app/components/buttons/Primary.vue
@@ -17,7 +17,7 @@ const base = 'font-inter flex justify-center items-center gap-3 font-inter font-
 const variant = computed(() => {
   if (props.variant === 'outlined') {
     return 'border-4 bg-transparent'
-  } return ''
+  } return 'border-4'
 })
 
 const color = computed(() => {
@@ -26,17 +26,17 @@ const color = computed(() => {
       if (props.variant === 'outlined') {
         return 'border-secondary text-black hover:bg-secondary'
       }
-      return 'bg-secondary hover:bg-secondary/80 text-black'
+      return 'border-secondary bg-secondary hover:bg-secondary/80 text-black'
     case 'primary':
       if (props.variant === 'outlined') {
         return 'border-primary text-primary hover:bg-primary hover:text-white'
       }
-      return 'bg-primary hover:bg-primary/80 text-white'
+      return 'border-primary bg-primary hover:bg-primary/80 text-white'
     case 'black':
       if (props.variant === 'outlined') {
         return 'border-black text-black hover:bg-black hover:text-white'
       }
-      return 'bg-black hover:bg-black/80 text-white'
+      return 'border-black bg-black hover:bg-black/80 text-white'
     default:
       return ''
   }