+++ app/layouts/default.vue
@@ -1,7 +1,9 @@
 <template>
+  <Header />
   <div>
     <slot />
   </div>
+  <Footer />
 </template>
 
 <script lang="ts" setup>